+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # CHANGELOG
 
+- __07/03/2023__
+  - Added `-Path` parameter, now both `-Path` and `-LiteralPath` parameters take `string[]` as input and support pipeline input.
+  - Added Pester tests, Code Coverage and coverage upload to [codecov.io](https://app.codecov.io/gh/santisq/PSTree).
+  - Removed `.Size` Property from `PSTreeFile` and `PSTreeDirectory` instances. The `Size` column has been renamed to `Length` and moved to the left-hand side of the `Hierarchy` column (I know it looks much better on the right-hand side :expressionless: but having it in the left allows for fixed width in the first 2 columns, which in turn brings less formatting issues :man_shrugging:...).
+
+    The default display for this column is available through `[PSTree.Internal._Format]::GetFormattedLength(...)`, for example:
+
+    ```powershell
+    Get-PSTree | Select-Object Hierarchy, @{ N='Size'; E={ [PSTree.Internal._Format]::GetFormattedLength($_.Length) }}
+    ```
+
+  - Added `GroupBy` tag to the default view, now trees are grouped by the source Path omg! Little example:
+
+    ```powershell
+    PS ..\PSTree> Get-PSTree .\src\, .\module\ -Exclude *\obj,*\bin
+
+      Source: C:\path\to\PSTree\src
+
+    Mode            Length Hierarchy
+    ----            ------ ---------
+    d----          0.00  B src
+    d----         10.30 KB └── PSTree
+    -a---        931.00  B     ├── ExceptionHelpers.cs
+    -a---        439.00  B     ├── PSTree.csproj
+    -a---          1.06 KB     ├── PSTreeDirectory.cs
+    -a---          4.01 KB     ├── PSTreeExtensions.cs
+    -a---        517.00  B     ├── PSTreeFile.cs
+    -a---        399.00  B     ├── PSTreeFileSystemInfo.cs
+    -a---          1.51 KB     ├── PSTreeFileSystemInfo_T.cs
+    -a---        897.00  B     ├── PSTreeHelper.cs
+    -a---        619.00  B     ├── PSTreeIndexer.cs
+    d----          1.13 KB     ├── Internal
+    -a---          1.13 KB     │   └── _Format.cs
+    d----          5.68 KB     └── Commands
+    -a---          5.68 KB         └── GetPSTreeCommand.cs
+
+      Source: C:\path\to\PSTree
+
+    Mode            Length Hierarchy
+    ----            ------ ---------
+    d----          6.22 KB module
+    -a---          1.54 KB ├── PSTree.Format.ps1xml
+    -a---          4.67 KB └── PSTree.psd1
+    ```
+
 - __03/22/2023__
   - `Get-PSTree` is now a binary cmdlet. Functionality remains the same. Big thanks to [SeeminglyScience](https://github.com/SeeminglyScience/) and [jborean93](https://github.com/jborean93/) for all their help!
   - Added `-Exclude` parameter to the cmdlet. The parameter accepts wildcards and patterns are matched with the object's `.FullName` property. For more details checkout [cmdlet docs](/docs/en-US/Get-PSTree.md).